+++ jeu.h
@@ -4,7 +4,7 @@
 #include <SDL2/SDL.h>
 #include "Game.h";
 
-int loadProjectilesMonstersDestrutiblesCoins(Game* jeu, int textures[], SDL_Renderer* ecran, SDL_Texture* textureTileSetV4, SDL_Texture* coin, SDL_Texture* heart, SDL_Texture* half);
-int load(Game* jeu, int textures[], SDL_Renderer* ecran, SDL_Texture* textureTileSetV4, SDL_Texture* textureTileSetWallsV2, SDL_Texture* key);
+short loadProjectilesMonstersDestrutiblesCoins(Game* jeu, short textures[], SDL_Renderer* ecran, SDL_Texture* textureTileSetV4, SDL_Texture* coin, SDL_Texture* heart, SDL_Texture* half);
+short load(Game* jeu, short textures[], SDL_Renderer* ecran, SDL_Texture* textureTileSetV4, SDL_Texture* textureTileSetWallsV2, SDL_Texture* key);
 
 #endif
